fix article edit page tag delete bug
In the old version code, when you click tag name in the article edit
page, the site will delete the tag directly, not the tag link.

In this update, the click action will delete the article-tag link
first, and then check for the tag link count, if the tag don't have
other link, the site will delete the tag also.